+++ src/main/java/com/asac/study_hub/controller/UserController.java
@@ -2,9 +2,9 @@ package com.asac.study_hub.controller;
 
 import com.asac.study_hub.controller.dto.common.BaseResponse;
 import com.asac.study_hub.controller.dto.common.SuccessType;
+import com.asac.study_hub.controller.dto.userDto.UserResponseDto;
 import com.asac.study_hub.controller.dto.userDto.signinDto.SigninRequestDto;
 import com.asac.study_hub.controller.dto.userDto.signupDto.SignupRequestDto;
-import com.asac.study_hub.controller.dto.userDto.signupDto.SignupResponseDto;
 import com.asac.study_hub.service.UserService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -14,9 +14,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
-
-import static org.springframework.http.HttpStatus.CREATED;
-import static org.springframework.http.HttpStatus.OK;
 
 @RestController
 @RequiredArgsConstructor
@@ -29,13 +26,13 @@ public class UserController {
 
     @CrossOrigin(origins = "http://localhost:5173")
     @PostMapping("/signup")
-    public BaseResponse<SignupResponseDto> signup(@RequestBody @Valid SignupRequestDto signupRequestDto) {
+    public BaseResponse<UserResponseDto> signup(@RequestBody @Valid SignupRequestDto signupRequestDto) {
         return BaseResponse.success(SuccessType.SIGNUP, userService.signup(signupRequestDto));
     }
 
     @CrossOrigin(origins = "http://localhost:5173", allowCredentials = "true")
     @PostMapping("/signin")
-    public BaseResponse<SignupResponseDto> signin(HttpServletRequest request, HttpServletResponse response, @RequestBody @Valid SigninRequestDto userDto) {
+    public BaseResponse<UserResponseDto> signin(HttpServletRequest request, HttpServletResponse response, @RequestBody @Valid SigninRequestDto userDto) {
         return BaseResponse.success(SuccessType.SIGNIN, userService.signin(request, response, userDto));
     }
 }
