+++ src/main/java/com/asac/study_hub/service/UserService.java
@@ -2,7 +2,7 @@ package com.asac.study_hub.service;
 
 import com.asac.study_hub.controller.dto.userDto.signinDto.SigninRequestDto;
 import com.asac.study_hub.controller.dto.userDto.signupDto.SignupRequestDto;
-import com.asac.study_hub.controller.dto.userDto.signupDto.SignupResponseDto;
+import com.asac.study_hub.controller.dto.userDto.UserResponseDto;
 import com.asac.study_hub.domain.User;
 import com.asac.study_hub.exception.CustomException;
 import com.asac.study_hub.exception.ExceptionType;
@@ -10,13 +10,10 @@ import com.asac.study_hub.repository.UserRepository;
 import com.asac.study_hub.util.SessionProvider;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
-import org.springframework.objenesis.strategy.SingleInstantiatorStrategy;
 import org.springframework.stereotype.Service;
 
 
@@ -27,7 +24,7 @@ public class UserService {
 
     UserRepository userRepository;
 
-    public SignupResponseDto signup(SignupRequestDto signupRequestDto) {
+    public UserResponseDto signup(SignupRequestDto signupRequestDto) {
         String email = signupRequestDto.getEmail();
         String name = signupRequestDto.getUserName();
 
@@ -41,18 +38,18 @@ public class UserService {
         signupRequestDto.setId(userRepository.findAll().size() + 1);
         User user = userRepository.save(SignupRequestDto.of(signupRequestDto));
 
-        return SignupResponseDto.builder()
+        return UserResponseDto.builder()
                 .userId(user.getId())
                 .status(HttpStatus.CREATED.value())
                 .build();
 
     }
 
-    public SignupResponseDto signin(HttpServletRequest request, HttpServletResponse response, SigninRequestDto userDto) {
+    public UserResponseDto signin(HttpServletRequest request, HttpServletResponse response, SigninRequestDto userDto) {
 
         saveSession(request, userDto);
         User user = userRepository.findByEmail(userDto.getEmail());
-        return SignupResponseDto.builder()
+        return UserResponseDto.builder()
                 .userId(user.getId())
                 .status(HttpStatus.OK.value())
                 .build();
